refactor(cli): nest commands under a2a parent command
Move agent-card, send, stream, task commands into commands/a2a/
directory and register them as subcommands of the `a2a` parent command.

CLI usage changes:
  a2x a2a agent-card <url>
  a2x a2a send <url> <message>
  a2x a2a stream <url> <message>
  a2x a2a task get <url> <task-id>
  a2x a2a task cancel <url> <task-id>
