+++ packages/cli/src/commands/a2a/send.ts
@@ -1,7 +1,7 @@
 import { Command } from 'commander';
 import crypto from 'node:crypto';
 import type { SendMessageParams } from 'a2x';
-import { printTask, printConnectionError, createClient } from '../format.js';
+import { printTask, printConnectionError, createClient } from '../../format.js';
 
 export const sendCommand = new Command('send')
   .description('Send a message to an A2A agent (blocking)')
