+++ packages/cli/src/commands/index.ts
@@ -1,4 +1,1 @@
-export { agentCardCommand } from './agent-card.js';
-export { sendCommand } from './send.js';
-export { streamCommand } from './stream.js';
-export { taskCommand } from './task.js';
+export { a2aCommand } from './a2a/index.js';
